==> source/0.EDA.ipynb <==
import os
import datetime
import IPython
import IPython.display
import matplotlib as mpl
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import seaborn as sns

%cd '/Users/benjochem/Desktop/Junior/Research'

sec_prices = pd.read_csv('Project/data/raw/vxx_security_prices.csv')
options = pd.read_csv('Project/data/raw/option price_vxx_5.2_010118-123119.csv')
volt = pd.read_csv('Project/data/raw/historical volatility_vxx.csv')
stand_options = pd.read_csv('Project/data/raw/standardized options_vxx.csv')

sec_prices

options[options.date == '12/27/2019']

stand_options[stand_options.date=='12/27/2019']

volt

sec_prices=sec_prices[['date','exchange_d','low','high','open','close','volume','return']]
