==> source/1.Treasury Preprocess.ipynb <==
# Treasury Preprocessing

Goal: Tansform Date column in the yields data to a common format that will be used in each data frame. Will need the same date format to properly join the data frames.

Yields data was sourced from the U.S. Department of the Treasury Daily Treasury Yield Curve Rates (https://www.treasury.gov/resource-center/data-chart-center/interest-rates/Pages/TextView.aspx?data=yield)

Will be used to inform the risk-free rate that corresponds in date and time to maturity for the options data (which will ultimately be used as an input to the multi-layer perceptron models)

import pandas as pd

%cd '/Users/benjochem/Desktop/Junior/Research'

yields = pd.read_excel('Project/data/raw/yields.xlsx')
yields

yields.isnull().sum()

# Convert date column to a format corresponding with options data (yyyymmdd)
#  yields will need to be matched to options by date
def date_to_numeric(date = []):
    converted = []
    for d in date:
        d = str(d)
        d = d[0:11]
        d = d.replace('-','')[0:8]
        d = int(d)
        converted.append(d)
    return converted

# apply function to Date column in yields and add correctly formatted date column
test = date_to_numeric(yields['Date'])
yields['date'] = test
yields.head()

# save correctly formated yields data to interim data directory
yields.drop('Date', axis=1).to_csv('Project/Data/interim/fixed_yields.csv', index = False)
