==> source/2.Options Preprocess.ipynb <==
# Options Preprocessing

Goal: Create a data frame containing all options and reduce it to only features that will be used in the model.
      Tansform date/exdate columns in the options data to a common format that will be used in each data frame. Will need the same date format to properly join the
      data frames.

The options data corresponds to options securities from the iPath S&P 500 VIX Short-Term Futures ETN (VXX).
The VXX was an ETN and thus traded like an ETF until its ultimate expiration on Jan. 31, 2019. Over its life the VXX was primarily used to speculate on and hedge against market volatility, and was never designed to be a buy-and-hold-investment. A similar project, the iPath Series B S&P 500 VIX Short Term Futures ETN (VXXB), was launched in 2018 with a longer maturity of 30 years to replace the VXX. It is in turn very similar to the VXX and thus this work may also be relevant to the VXXB.

Note: The VIX is a market index that measures expected price swings in the S&P 500 Index over the next 30 days, based on the trading of options contracts linked to it and is often called a fear gague for the market as a whole. VXX / VXXB provides exposure to futures contracts on the VIX.

import pandas as pd

%cd '/Users/benjochem/Desktop/Junior/Research'

#read in all options files (range from 2010 to 2019)
opt5=pd.read_csv('Project/data/raw/option price_vxx_5.2_010118-123119.csv')
opt4=pd.read_csv('Project/data/raw/option price_vxx_4.2_010116-123117.csv')
opt3=pd.read_csv('Project/data/raw/option price_vxx_3.2_010114-123115.csv')
opt2=pd.read_csv('Project/data/raw/option price_vxx_2.2_010112-123113.csv')
opt1=pd.read_csv('Project/data/raw/option price_vxx_1.2_010110-123111.csv')

# concatenate options data by 'stacking' subsequent data frames on top of one another
# oldest option at index 0
options = pd.concat([opt1,opt2,opt3,opt4,opt5])
print(len(options))
options.head()

options.isnull().sum()

#convert data frame of all options to a csv file and save in interim data directory
options.to_csv('Project/data/interim/full_options.csv', index = False)

options=pd.read_csv('Project/data/interim/full_options.csv')
print(len(options))
options.head()

#reduce data frame to features to be used in the model
options=options[['date','exdate','cp_flag','strike_price','volume','open_interest','delta','gamma','impl_volatility','best_bid','best_offer']]
options.head()

#function to match data structure of the options dates to treasury data dates (yyyymmdd)
def date_to_numeric(date = []):
    converted = []
    for d in date:
        d = d.strip().split('/')
        day,month,year = d[0], d[1], d[2]
    
        # leading zeros on days/months
        if len(day) == 1: 
            day = '0' + day
        if len(month) == 1:
            month = '0' + month

        string = year + day + month
        converted.append(string)
    
    return converted

# apply function to date/exdate columns in options and add correctly formatted Date/exDate column
options['Date'] = date_to_numeric(options.date)
options['exDate'] = date_to_numeric(options.exdate)
options.head()

#drop old date columns
options=options.drop('date',axis=1)
options=options.drop('exdate',axis=1)

# new full options dataframe with Dates and exDates in correct format
print(len(options))
options.head()

#convert data frame of all options with correct dates to a csv file and save in interim data directory
options.to_csv('Project/data/interim/full_options_w_dates.csv', index = False)
