==> source/3.Options Feature Engineering.ipynb <==
# Options Feature Engineering

Goal: Time to expiration is a common feature used in options pricing. Must derive time to expiration for each option in our data by finding difference between the current date and expiration date

Note: In general longer time to expiration results in higher option price because there is more time for the option to finish "in-the-money"

import pandas as pd
from datetime import datetime, timedelta

%cd '/Users/benjochem/Desktop/Junior/Research'

options = pd.read_csv('Project/data/interim/full_options_w_dates.csv')
print(len(options))
options.head()

# find the time-to-expiration for each option by finding the difference between exDate and Date of our model
# create a function to find the time-to-expiration (TTE)
def find_expiration(date = [], exdate = []):
    TTE =[]
    
    for i in range(0, len(date)):
        S = str(date[i])
        E = str(exdate[i])
        
        S_year, S_month, S_day = S[0:4], S[4:6], S[6:8]
        E_year, E_month, E_day = E[0:4], E[4:6], E[6:8]
        
        start = datetime(year = int(S_year), month = int(S_month), day = int(S_day))
        end = datetime(year = int(E_year), month = int(E_month), day = int(E_day))
        
        diff = end - start
        TTE.append(diff)
    
    return TTE

options['TTE'] = find_expiration(options.Date, options.exDate)
options.head()

# used to covert timestamp time to expiration to integer number of days
def numeric_maturity(TTE = []):
    converted =[]
    for i in TTE:
        i = str(i).split(' ')
        day = int(i[0])
        converted.append(day)
    return converted

options['TTE'] = numeric_maturity(options.TTE)
options.head()

print(len(options))
options.isnull().sum()

options.to_csv('Project/data/interim/full_options_w_dates_TTE.csv', index = False)
