==> source/4.Securities Feature Engineering.ipynb <==
# Security Feature Engineering

Goal: Reduce the securities data to only features that will be used in our model. Tansform date column in the securities data to a common format that will be used in each data frame. Will need the same date format to properly join the data frames.

Create a derived implied volatility feature for each date. Standard deviation of the last 20 days of stock close prices. Will be used as an additional input to the model.

import pandas as pd
import numpy as np

%cd '/Users/benjochem/Desktop/Junior/Research'

securities = pd.read_csv('Project/data/raw/vxx_security_prices.csv')
securities = securities [['date','close','volume','return']]
securities.head()

#function to match data structure of the securities dates to treasury data / options data dates (yyyymmdd)
def date_to_numeric(date = []):
    converted = []
    for d in date:
        d = d.strip().split('/')
        day,month,year = d[0], d[1], d[2]
    
        # leading zeros on days/months
        if len(day) == 1: 
            day = '0' + day
        if len(month) == 1:
            month = '0' + month

        string = year + day + month
        converted.append(string)
    
    return converted

securities['Date'] = date_to_numeric(securities.date)
securities.head()

securities.drop(labels = 'date', axis = 1, inplace = True)
securities.head()

securities

# Volatility over life of the option is assumed to be historical volatility of the last 20 trading days
# rolling essentially aggregates the last 20 observations and requires some function be applied 
# lambda function calculates the standard deviation of day to day price changes for the last 20 days

securities['vol_20'] = securities.close.rolling(20).apply(lambda x: (np.diff(x) / x[:-1]).std())

securities.to_csv('Project/data/interim/securities_w_date.csv', index = False)

securities = pd.read_csv('Project/data/interim/securities_w_date.csv')
