==> source/5. Create Final Data.ipynb <==
# Final Data

Goal: Combine yields, securities and options data to create a single data set of features and labels used in the multilayer perceptron model

import pandas as pd

%cd '/Users/benjochem/Desktop/Junior/Research'

yields = pd.read_csv('Project/data/interim/fixed_yields.csv')
options = pd.read_csv('Project/data/interim/full_options_w_dates_TTE.csv')
securities = pd.read_csv('Project/data/interim/securities_w_date.csv')

yields.head()

options.head()

securities.head()

#find number of unique dates in options
#options with the same date will be associated with the same row in the treasury yields data 
dates_in_options=set(options.Date.values)

#find options that do not have a corresponding row for treasury yields at their date
not_in_treasury=dates_in_options - set(yields.date.values)

# all options dates that do not have corresponding treasury yields
not_in_treasury

#create a dictionary from the treasury yields
# date as key, dictionary of the different rates(based on maturity) as value
#each date in yields will be a nested dictionary in yields_dictionary
yields_dictionary = {}
for i, row in yields.iterrows(): 
    d = dict(row)                                 # create a dictionary of the current row in the yields data
    yields_dictionary[int(d['date'])] = d         # add dictionary d to yields_dictionary with the date of the row as the key and the entire row as the value
    del yields_dictionary[int(d['date'])]['date'] # drop date from the nested dictionary value since it is used as the key

#create a set of all dates in yields data
yield_dates= set(yields.date.values)

#create data frame that includes ONLY options that have yield data
options_w_yields=options[options.Date.isin(yield_dates)]

# how many total options have corresponding yields data
len(options_w_yields)

def find_t_yield(row):
    
        date = row.Date #date value for the option on current row
        tte=row.TTE #time to expiration for the option on current row
        # detrmine which t-yield rate to use based on TTE
        
        if tte <= 45:
            length = '1 Mo'
        elif tte <= 75:
            length = '2 Mo'
        elif tte <= 135:
            length = '3 Mo'
        elif tte <= 270:
            length = '6 Mo'
        elif tte <= 547:
            length = '1 Yr'
        elif tte <= 912:
            length = '2 Yr'
        elif tte <= 1460:
            length = '3 Yr'
        elif tte <= 2190:
            length = '5 Yr'
        elif tte <= 3102:
            length = '7 Yr'
        elif tte <= 3975:
            length= '10 Yr'
        elif tte <= 6625:
            length = '20 Yr'
        else:
            length = '30 Yr'
            
        rate = yields_dictionary[date][length] #find and select the correct rate in date yields dictionary (go to correct date first then go to correct rate becasue it is nested)
     
        if rate >= 0: #2 month rates are missing for most of options (didnt start until 2018) so use 3 month rate if rate is not found or nan
            pass
        else:
            rate = yields_dictionary[date]['3 Mo']
    
        return rate 

options_w_yields['treasury_rate'] = options_w_yields.apply(lambda row: find_t_yield(row), axis=1)
options_w_yields

#re-order columns to be used
options_w_yields=options_w_yields[['Date','exDate','cp_flag','strike_price','volume','treasury_rate','TTE','open_interest','best_bid','best_offer']]
options_w_yields

securities=pd.read_csv('Project/data/interim/securities_w_date.csv')
securities

# drop securities with duplicate dates. Will result in having duplicate rows in final features data
print(len(securities))
securities=securities.drop_duplicates(subset='Date')
print(len(securities))

features = options_w_yields.merge(securities, on = 'Date', how = 'left')

print(len(features))
features.head()

features = features.rename({'close':'sec_price','volume_y':'sec_volume','return':'sec_return','volume_x':'option_volume'}, axis = 1)
features.head()

features = features[['Date','exDate','cp_flag','TTE','strike_price','treasury_rate','sec_price','option_volume','open_interest','sec_volume','sec_return','vol_20','best_bid','best_offer']]
features.head()

features.to_csv('Project/data/interim/features.csv', index = False)
